chore: Refactor library to use smaller modules
The `index.js` file got pretty big over time. This commit splits it into smaller modules:
- The core logic remains in `index.js`
- The `munge` functions are moved to the `lib/munge` directory. Each module detects and wraps one kind of target.
- Helper functions are moved out of `index.js` to the `lib` directory

Using multiple modules instead of a single `index.js` enables easier future extensions.
